move open_raw and open_converted to echodata/open_file.py
