+++ echopype/__init__.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import, division, print_function
 from .convert import Convert
 from .process import Process
-from .echodata import open_converted
+from .echodata import open_converted, open_raw
 from .calibrate import calibrate_func as calibrate
 
 from _echopype_version import version as __version__
