+++ echopype/echodata/__init__.py
@@ -2,4 +2,5 @@
 EchoData is an object that handles interfacing raw converted data.
 It is used for calibration and other processing.
 """
-from .echodata import EchoData, open_converted
+from .echodata import EchoData
+from .open_file import open_raw, open_converted
