+++ echopype/tests/test_convert_azfp.py
@@ -11,7 +11,7 @@ import numpy as np
 import xarray as xr
 import pandas as pd
 from scipy.io import loadmat
-from ..convert import open_raw
+from ..echodata import open_raw
 
 
 azfp_path = Path('./echopype/test_data/azfp')
