+++ echopype/tests/test_convert_ek80.py
@@ -4,7 +4,7 @@ import numpy as np
 import xarray as xr
 import pandas as pd
 from scipy.io import loadmat
-from ..convert import open_raw
+from ..echodata import open_raw
 
 ek80_path = Path('./echopype/test_data/ek80/')
